Refactored some code, mostly controllers

+++ server/src/controllers/bookmark.controller.js
@@ -1,4 +1,5 @@
 import { Bookmark } from "../models/bookmark.model.js";
+import { User } from "../models/user.model.js";
 
 const createBookmark = async (req, res) => {
     try {
@@ -13,8 +14,14 @@ const createBookmark = async (req, res) => {
             return res.status(400).json({ message: "Movie Type can only be 'movie' or 'series'" });
         }
 
+        const existingUser = await User.findById(userId);
+
+        if (!existingUser) {
+            return res.status(404).json({ message: "User does not exist" });
+        }
+
         const existingBookmark = await Bookmark.findOne({
-            userId,
+            user: userId,
             imdbMovieId: movieId,
         });
 
@@ -31,10 +38,7 @@ const createBookmark = async (req, res) => {
 
         return res.status(201).json({
             message: "Bookmark successful",
-            bookmark: {
-                userId,
-                imdbMovieId,
-            }
+            bookmark,
         })
     } catch (error) {
         return res.status(500).json({ message: "Internal server error", error: error.message });
@@ -58,15 +62,22 @@ const getAllBookmarks = async (req, res) => {
     }
 }
 
-const getBookmarksByUserId = async (req, res) => {
+const getAllBookmarksByUserId = async (req, res) => {
     try {
         const { userId } = req.params;
-        const bookmarks = await Bookmarks.find({
-            userId
+        const existingUser = await User.findById(userId);
+
+        if (!existingUser) {
+            return res.status(404).json({ message: "User does not exist" });
+        }
+
+        const bookmarks = await Bookmark.find({
+            user: userId
         });
 
+
         if (bookmarks.length === 0) {
-            return res.status(404).json({ message: "User has no bookmarks" });
+            return res.status(404).json({ message: "User has no bookmarks or User does not exist" });
         }
 
         return res.status(200).json({
@@ -78,24 +89,29 @@ const getBookmarksByUserId = async (req, res) => {
     }
 }
 
-const deleteBookmarkForUser = async (req, res) => {
+const deleteMovieBookmarkByUserId = async (req, res) => {
     try {
         const { userId, movieId } = req.params;
+        const existingUser = await User.findById(userId);
+
+        if (!existingUser) {
+            return res.statsu(404).json({ message: "User does not exist" });
+        }
 
         const deletedBookmark = await Bookmark.deleteOne({
-            userId,
+            user: userId,
             imdbMovieId: movieId,
         });
 
         if (deletedBookmark.deletedCount === 0) {
-            return res.status(400).json({ message: "Bookmark could not be found for user, therefore no bookmark was deleted" });
+            return res.status(400).json({ message: "Bookmark does not exit" });
         }
 
         return res.status(200).json({
             message: "Bookmark deleted",
             movieDeletedInfo: {
                 userId,
-                imdbMovieId,
+                movieId,
             }
         });
     } catch (error) {
@@ -108,7 +124,7 @@ export {
     createBookmark,
     // GET
     getAllBookmarks,
-    getBookmarksByUserId,
+    getAllBookmarksByUserId,
     // DELETE
-    deleteBookmarkForUser,
+    deleteMovieBookmarkByUserId,
 }
